Separate SeriesController class into partial files
